% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCfunctions.R
\name{auc}
\alias{auc}
\title{Extracts auc (se) and ROC curve estimates}
\usage{
auc(x1 = 20, x2 = NULL, fit, model = "PH", ROC = FALSE, ngrid = 100)
}
\arguments{
\item{x1}{a value of covariate x1 at which to estimate auc and ROC curve. Can be absent.}

\item{x2}{a value of covariate x2 at which to estimate auc and ROC curve. Can be absent.}

\item{fit}{a coxph/croc object of model fit.}

\item{model}{a character string specifying the model for the conave ROC curve, one of PH and PrevH. Default PH.}

\item{ROC}{logical, if TRUE, ROC estimates are produced in addition to AUC. Default is FALSE.}

\item{ngrid}{number of equal-spaced FPR values on unit interval to compute TRP. Dafault 100.}
}
\value{
either a vector of auc estimates or (if ROC = TRUE) a list of a vector of auc estimates and dataframe of FPR and TPR.
}
\description{
This function extracts auc (se) and ROC curve estimates from PHroc and PrevHroc model fits
}
\details{
The proportional hazards ROC (PHroc) model assumes proportional hazards between healthy and diseased
populations: \eqn{h(y|d=1,x)/h(y|d=0,x) = \theta} and produces ROC curve \eqn{TPR = FPR^\theta} and AUC \eqn{1/(1+\theta)}.
With covariate, PHroc model takes the form of \eqn{h(y) = h_0(y) exp(b1 d + b2 x + b3 d*x)} where y, d, and x are score, status,
and covariate, respectively, implying \eqn{h(y|d=1,x)/h(y|d=0,x) = exp(b1 + b3 x)}. For binary x, \eqn{\theta = exp(b1)} and \eqn{exp(b1+b3)}
at \eqn{x = 0} and \eqn{x = 1}, respectively. The proportional reversed hazards ROC (PrevHroc) assumes proportional reversed hazards
instead, produces ROC curve \eqn{TPR = 1-(1-FPR)^{1/\theta}}, and shares the same AUC \eqn{1/(1+\theta)}.
With \eqn{\lambda(y)} as the reversed hazard function, PrevHroc model takes the form of \eqn{\lambda(y) = \lambda_0(y) exp(b1 d + b2 x + b3 d*x)},
implying \eqn{\lambda(y|d=1,x)/\lambda(y|d=0,x) = exp(b1 + b3 x)}. Same as in PHroc, \eqn{\theta = exp(b1)} and \eqn{exp(b1+b3)}
at \eqn{x = 0} and \eqn{x = 1}.

Delta method is used to derive s.e. of auc.
}
\examples{
auc(fit,"PH")
auc(x1=1,x2=.6,fit,"PrevH",ROC=TRUE)
}
\references{
#' @export
}
\author{
Zhen Chen (zhen.chen@nih.gov)
}
