% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bin2lonlat}
\alias{bin2lonlat}
\title{Longitude and latitude from bin number.}
\usage{
bin2lonlat(bins, nrows)
}
\arguments{
\item{bins}{bin number}

\item{nrows}{number of rows in this grid}
}
\description{
Generate longitude and latitude coordinates from bin number.
}

