% Generated by roxygen2 (4.0.2): do not edit by hand
\name{binmap}
\alias{binmap}
\title{Bin map}
\usage{
binmap(bin, ras, init = NULL)
}
\arguments{
\item{bin}{bin number}

\item{ras}{RasterLayer}

\item{init}{optional initial values for bin structure}
}
\description{
mapping between bins and a given raster
}

