% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chla_compute.R
\name{chla_compute}
\alias{chla_compute}
\title{MODISA chlorophyll-a in the southern hemisphere.}
\usage{
chla_compute(days, maxbin = NULL)
}
\arguments{
\item{days}{dates to calculate}

\item{maxbin}{a maximum bin to use, for specialist use only}
}
\value{
tibble of bins and chlorophyll-a values
}
\description{
Compute chlorophyll data in the L3 bins, using both Johnson and Oceancolor algorithms.
}
\details{
Use `raadtools::ocfiles(product = "MODISA", varname = "RRS", type = "L3b")` to get all daily MODISA L3 bin RRS files, and
explore the `date` column to generate a set of `days` for this function.
}
