% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croc-package.R
\docType{package}
\name{croc-package}
\alias{croc}
\alias{croc-package}
\title{croc: Tools for Ocean Colour}
\description{
Work with data from NASA's ocean colour data, primarily the 'Level-3 bin'
files. There are functions to read and work with the ragged array 'L3 bins'
mesh and calculate biophysical properties from remote sensing reflectance data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sosoc/croc}
  \item Report bugs at \url{https://github.com/sosoc/croc/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Norman Kuring (Author of L3bin code (NASA)) [contributor]
  \item Robert Johnson (Developed southern ocean algorithm for chlorophyll) [contributor]
}

}
\keyword{internal}
