% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crocfuncollection.R
\name{croc.default}
\alias{croc.default}
\title{Estimate concave ROC curves and associated summary measures}
\usage{
\method{croc}{default}(x, y, model, ...)
}
\arguments{
\item{x}{a a matix of predictors. This should not include an intercept and already contains interactions and so on.}

\item{y}{a numeric vector containing the outcome (no censoring for now).}

\item{model}{a character string specifying the model for the conave ROC curve, one of PH, PrevH, PBN, Emprical.}
}
\value{
an object of class coxph representing the fit.
}
\description{
This function estimates concave curves and various summary measures.
}
