% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crocfuncollection.R
\name{croc.formula}
\alias{croc.formula}
\title{The formula wrapper for croc.default}
\usage{
\method{croc}{formula}(formula, data, model = c("PH", "PrevH"), subset, weights, na.action, ...)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~ operator, and the terms on the right.}

\item{data}{a data.frame in which to interpret the variables named in the formula, or in the subset and the weights argument.}

\item{model}{a character string specifying the model for the conave ROC curve, one of PH, PrevH, PBN, Emprical.}

\item{subset}{an expression indicating which subset of the rows of data should be used in the fit. All observations are included by default.}

\item{weights}{a vector of case weights.}
}
\value{
an object of class coxph and croc representing the fit.
}
\description{
This function provides the formula wrapper for croc.default function.
}
