% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crop.R
\name{crop_init}
\alias{crop_init}
\title{Crop L3 init object with an extent}
\usage{
crop_init(x, ext)
}
\arguments{
\item{x}{initbin object}

\item{ext}{raster extent object, or object to create an extent from}
}
\value{
integer vector of bins
}
\description{
Crop L3 list, returns bins that fall within the extent.
}
\examples{
\dontrun{ init <- initbin(24)
crop_init(init, extent(100, 110, -50, -45))
}
}

