% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crop.R
\name{densify}
\alias{densify}
\title{Insert redundant vertices to allow reprojection.}
\usage{
densify(x, maxdist, longlat = longlat)
}
\arguments{
\item{x}{2 row matrix of coordinates to densify}

\item{maxdist}{the maximum distance}

\item{longlat}{use longlat on the ellipsoid for distance?}
}
\description{
Densify objects.
}

