% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bins.R
\name{initbin}
\alias{initbin}
\title{Initialize values for a particular binning}
\usage{
initbin(NUMROWS = 2160)
}
\arguments{
\item{NUMROWS}{relevant number of L3 bin rows}
}
\description{
Set up the basic values for the bin scheme for given number of rows.
}

