% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lonlat2bin}
\alias{lonlat2bin}
\title{Generate bin number from longitude latitude.}
\usage{
lonlat2bin(lon, lat, NUMROWS)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{NUMROWS}{number of rows}
}
\description{
Bin number from longitude and latitude for a given grid with NUMROWS unique latitudes.
}

