% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lonlatFromCell}
\alias{lonlatFromCell}
\title{Raster cell longitude and latitudes}
\usage{
lonlatFromCell(object, cell = NULL, spatial = FALSE)
}
\arguments{
\item{object}{a raster object}

\item{cell}{the cell numbers}

\item{spatial}{return locations as SpatialPoints object instead of a matrix.}
}
\value{
the lon,lat locations for the requested cells.
}
\description{
Extract longitude and latitude of raster cells.
}
\details{
Extract the longitude and latitude of the center of the requested
cells of a Raster* object, similar to \code{xyFromCell}.

from SGAT
}

