% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid.R
\name{oceandata-grids}
\alias{oceandata-grids}
\alias{bin_smi}
\title{Standard Mapped Image}
\usage{
bin_smi(bins, value, platform = "MODISA")
}
\arguments{
\item{bins}{bin index values}

\item{value}{values with bins}

\item{platform}{name of platform, e.g. MODISA, SeaWiFS}
}
\description{
SMI is the Standard Mapped Image, a longitude-latitude affine grid
at zonally-native resolution. See details.
}
\details{
SeaWiFS is 2160 rows for 9.2 km resolution
MODISA is 4320 rows for 4.6 km resolution (also provided in 9.2km)
VIIRS is 4320 rows like MODIS. 
The zonal spread of the L3 bins is applied to the longitude-latitude 
regular grid by _a simple method that changed slightly in 2017 ..._
is trivial but inherently lossy. The initial focus of `roc` on the L3 bins was
precisely to put focus on the statistical fidelity features of the L3 bins, as
per Technical Report No. 32.

(There's no reason to differentiate MODISA and MODIST in terms of the grid it's 
just that this token was chosen to distinguish from SeaWiFS in the first versions,
 and this might be generalized in future).
}
