% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crocfuncollection.R
\name{ph.fit}
\alias{ph.fit}
\title{Fit Proportional Hazards (PH) model}
\usage{
ph.fit(x, y, ...)
}
\arguments{
\item{x}{a matix of predictors. This should not include an intercept.}

\item{y}{a numeric vector containing the outcome (no censoring for now).}
}
\value{
an object of class coxph representing the fit.
}
\description{
This function fits the PH model given outcome y and covariates x
}
