% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crocfuncollection.R
\name{print.croc}
\alias{print.croc}
\title{Print method for class "croc"}
\usage{
\method{print}{croc}(x, digits = max(1L, getOption("digits") - 3L), signif.stars = FALSE, ...)
}
\arguments{
\item{x}{an object of class "croc".}

\item{digits}{the number of significant digits to use when printing.}

\item{signif.stars}{logical. If TRUE, ‘significance stars’ are printed for each coefficient.}

\item{getOption}{????.}
}
\value{
A list of statistics????.
}
\description{
This function provides the print method for class "croc".
}
