% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readL3.R
\name{read_l3}
\alias{read_l3}
\title{L3 bin from NetCDF 4 compound types.}
\usage{
read_l3(date, ..., inputfiles)
}
\arguments{
\item{date}{date to read}

\item{...}{ignored}

\item{inputfiles}{input file dataframe}
}
\value{
tibble
}
\description{
WIP
}
\examples{
root <- "data/oceandata.sci.gsfc.nasa.gov/SeaWiFS/L3BIN"
data_dir <- getOption("default.datadir")  ## raadtools 
files <- structure(list(file = file.path(root, "1997/247/S1997247.L3b_DAY_RRS.nc"), 
fullname = file.path(data_dir, file.path(root, "1997/247/S1997247.L3b_DAY_RRS.nc")), 
date = structure(873331200, class = c("POSIXct", "POSIXt"
), tzone = "GMT")), .Names = c("file", "fullname", "date"
), row.names = 1L, class = "data.frame")
read_l3(inputfiles = files)

}
