% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readL2.R
\name{readL2}
\alias{readL2}
\title{Read L2 ocean colour files}
\usage{
readL2(file, vartype = c("swath", "meta"), data.frame = TRUE,
  sdsonly = FALSE)
}
\arguments{
\item{file}{L2 ocean colour file}

\item{vartype}{swath or meta}

\item{data.frame}{return a data.frame or the raster}

\item{sdsonly}{return only SDS names}
}
\description{
Read L2 ocean colour files
}

