% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readL3.R
\name{readL3}
\alias{readL3}
\title{Basic L3 bin files}
\usage{
readL3(x, vname, bins = TRUE, check = TRUE)
}
\arguments{
\item{x}{filename path to L3 bin OC file (HDF4)}

\item{vname}{names of VData parameters to read (will read both _sum and _ssq)}

\item{bins}{read out the bin number and other metadata (default TRUE)}

\item{check}{test for presence of auxiliary files, TRUE or FALSE}
}
\description{
Read from L3 bin.
}

