% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readL3img.R
\name{readL3img}
\alias{readL3img}
\title{Read NASA ocean colour browser images (PNG format)}
\usage{
readL3img(date, platform = c(A = "MODISA", S = "SeaWiFS"), tres = c(DAY =
  "", `8D` = "8 days", MO = "1 month", YR = "1 year"), varname = c(CHL =
  "chlor_a"))
}
\arguments{
\item{date}{date of image to find}

\item{platform}{which satellite platform (currently MODISA and SeaWiFS)}

\item{tres}{temporal resolution to read}

\item{varname}{name of variable (currently chlorophyll-a)}
}
\description{
Read ocean colour imagery from the NASA site, files will be downloaded and read in with
\code{brick}. If a file with an identical name is already present in the current working directory the download
is skipped.
}

