% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid.R
\name{regridder}
\alias{regridder}
\title{Regrid ocean colour.}
\usage{
regridder(tgrid, file, agg = FALSE, fun = mean, NROWS)
}
\arguments{
\item{tgrid}{target grid}

\item{file}{L3 bin file name with raw RRS wavelengths}

\item{agg}{aggregate values?}

\item{fun}{function to aggregate by if agg is \code{TRUE}}

\item{NROWS}{number of L3 rows}
}
\description{
Regrid NASA ocean colour RRS values to standard Mapped image with the Johnson chl-a algorithm.
}
