% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{roc-package}
\alias{roc-package}
\alias{roc-package,}
\alias{roc}
\title{R tools for ocean colour.}
\description{
Read data from Ocean Color Level-3 bin files.
}
\details{
\tabular{ll}{
Package: \tab roc\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2014-08-28\cr
License: \tab GPL-3 \cr
}
}
\references{
\url{https://github.com/mdsumner}, \url{http://www.hdfgroup.org/}, \url{http://oceandata.sci.gsfc.nasa.gov/}
}
\author{
Michael D. Sumner \cr \url{https://github.com/mdsumner} \cr \cr
Maintainer: Michael D. Sumner \email{mdsumner@gmail.com} \cr
}
\keyword{package}
