% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins.R
\name{sinu_coords}
\alias{sinu_coords}
\title{Sinusoidal coordinates from bin number}
\usage{
sinu_coords(bin, NUMROWS)
}
\arguments{
\item{bin}{bin L3 bin number}

\item{NUMROWS}{NUMROWS grid spec for these bins}
}
\description{
Sinusoidal coordinates from bin number
}
