% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vdatainfo}
\alias{vdatainfo}
\title{Vdata names and field names in HDF4.}
\usage{
vdatainfo(file)
}
\arguments{
\item{file}{name of HDF4 file containing Vdata}
}
\value{
Named list, each element with the field names of the named Vdata.
}
\description{
Retrieve vdata names and their field names from HDF4.
}
\examples{
\dontrun{
f <- "A2014202.L3b_DAY_PIC.main"
if (!file.exists(f)) {
  srcf <- sprintf("http://oceandata.sci.gsfc.nasa.gov/cgi/getfile/\%s.bz2", f)
  download.file(srcf, sprintf("\%s.bz2", f), mode = "wb")
  system(sprintf("bunzip2 \%s.bz2", f))
}
  vdatainfo(f)


}
}

