% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCfunctions.R
\name{auc}
\alias{auc}
\title{This function extracts auc (se) and ROC curve estimates from PH and PrevH model fits with <= 2 covariates}
\usage{
auc(x1 = 20, x2 = NULL, fit, ROC = FALSE, ngrid = 100)
}
\arguments{
\item{x1}{a value of covariate x1 at which to estimate auc and ROC curve. Can be absent.}

\item{x2}{a value of covariate x2 at which to estimate auc and ROC curve. Can be absent.}

\item{fit}{a coxph/croc object of model fit.}

\item{ROC}{logical, if TRUE, ROC estimates are produced in addition to AUC. Default is FALSE.}

\item{ngrid}{number of equal-spaced FPR values on unit interval to compute TRP.}
}
\value{
either a vector of auc estimates or (if ROC = TRUE) a list of a vector of auc estimates and dataframe of FPR and TPR.
}
\description{
This function extracts auc (se) and ROC curve estimates from PH and PrevH model fits with <= 2 covariates
}
