% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bins.R
\name{bin2bounds}
\alias{bin2bounds}
\title{Calculate bin boundaries from bin number}
\usage{
bin2bounds(bin, NUMROWS)
}
\arguments{
\item{bin}{bin number}

\item{NUMROWS}{relevant number of L3 bin rows}
}
\description{
Calculate bin boundaries from bin number
}

