% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bins.R
\name{binmap}
\alias{binmap}
\title{Bin map}
\usage{
binmap(bin, ras, init = NULL)
}
\arguments{
\item{bin}{bin number}

\item{ras}{RasterLayer}

\item{init}{optional initial values for bin structure}
}
\description{
mapping between bins and a given raster
}

