% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{chla}
\alias{chla}
\title{Estimate chlorophyll-a from NASA ocean colour.}
\usage{
chla(x, sensor, algo = c("oceancolor", "johnson"), quiet = TRUE, ...)
}
\arguments{
\item{x}{list object with Remote Sensing Reflectance wavelengths (see Details)}

\item{sensor}{which satellite sensor}

\item{algo}{algorithm to use, oceancolor or Johnson et al. (2013)}

\item{quiet}{logical to emit warnings or not}

\item{...}{not used}
}
\description{
Estimate chlorophyll-a from Remote Sensing Reflectance wavelengths. Use SeaWiFS or MODISA with original or Johnson algorithm
for the Southern Ocean.
}
\examples{
#f <- "S1998001.L3b_DAY_RRS.main"
#x <- readL3(f)
#asub <- x$bin_num < initlist()$totbins / 2
#ll <- bin2lonlat(x$bin_num[asub])
#sw <- swchl(x)[asub]
#js <- swchl(x, johnson = TRUE)[asub]
#par(mfrow = c(2,1))
#plot(ll, col = palr::chlPal(sw), pch = ".")
#plot(ll, col = palr::chlPal(js), pch = ".")

## setup a polar raster
}
