% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc.R
\name{chla}
\alias{chla}
\title{Estimate chlorophyll-a from NASA ocean colour.}
\usage{
chla(x, sensor, algo = c("oceancolor", "johnson"))
}
\arguments{
\item{x}{list object with Remote Sensing Reflectance wavelengths (see Details)}

\item{sensor}{which satellite sensor}

\item{algo}{algorithm to use, oceancolor or Johnson et al. (2013)}
}
\description{
Estimate chlorophyll-a from Remote Sensing Reflectance wavelengths. Use SeaWiFS or MODISA with original or Johnson algorithm
for the Southern Ocean.
}
\examples{
\dontrun{
f <- "S1998001.L3b_DAY_RRS.main"
x <- readL3(f)
asub <- x$bin_num < initlist()$totbins / 2
ll <- bin2lonlat(x$bin_num[asub])
sw <- swchl(x)[asub]
js <- swchl(x, johnson = TRUE)[asub]
par(mfrow = c(2,1))
plot(ll, col = raadtools::chl.pal(sw), pch = ".")
plot(ll, col = raadtools::chl.pal(js), pch = ".")

## setup a polar raster
require(raster)
require(rgdal)
prj <- "+proj=laea +lat_0=-90 +lon_0=147 +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"
p <- project(cbind(ll[[1]], ll[[2]]), prj)
r <- raster(extent(p), crs = prj, res = c(72000, 72000))
sr <- rasterize(p, r, field = sw, fun = mean)
jr <- rasterize(p, r, field = js, fun = mean)
plot(brick(sr, jr), col = pal$cols, breaks = pal$breaks, legend = FALSE)
}
}

