% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readoc.r
\name{readoc}
\alias{readoc}
\title{Read and summarize ocean colour variables between a date range.}
\usage{
readoc(daterange, xylim = NULL, varname = "CHL_RJ", grid = NULL,
  platform = "MODISA", ...)
}
\arguments{
\item{daterange}{two dates, specifying a start and end day}

\item{xylim}{extent specification (in longitude/latitude)}

\item{varname}{variable to read, defaults to Johnson 2013 chlorophyll-a}

\item{grid}{grid specification to rasterize to (optional)}

\item{platform}{which satellite source (MODSIA, SeaWiFS, ...)}

\item{...}{unused}
}
\description{
Read and summarize ocean colour variables.
}

