% Generated by roxygen2 (4.0.2): do not edit by hand
\name{regridder}
\alias{regridder}
\title{Regrid ocean colour.}
\usage{
regridder(tgrid, file, agg = FALSE, fun = mean)
}
\arguments{
\item{tgrid}{target grid}

\item{file}{L3 bin file name with raw RRS wavelengths}
}
\description{
Regrid NASA ocean colour RRS values to standard Mapped image with the Johnson chl-a algorithm.
}

